requirements.txt
asgiref==3.8.1
chardet==5.2.0
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.1
django-cors-headers==4.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.58.5
fpdf2==2.8.3
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
reportlab==4.4.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
